+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Use `name` and `summary` attributes to create post title.
 - Added `preview_url` attribute to attachments in Mastodon API.
 - Added API endpoint for cancelling invoices.
+- Added optional `chain_metadata.description` field to Monero blockchain config.
 
 ### Changed